+++ example/distance-hc-sr04.js
@@ -1,0 +1,33 @@
+'use strict';
+
+var Gpio = require('../').Gpio,
+  trigger = new Gpio(23, {mode: Gpio.OUTPUT}),
+  echo = new Gpio(22, {mode: Gpio.INPUT, alert: true});
+
+// The number of microseconds it take sound to travel 1cm at 20 degrees celcius
+var MICROSECDONDS_PER_CM = 1e6/34321;
+
+trigger.digitalWrite(0); // Make sure trigger is low
+
+(function () {
+  var startTick;
+
+  echo.on('alert', function (level, tick) {
+    var endTick,
+      diff;
+
+    if (level == 1) {
+      startTick = tick;
+    } else {
+      endTick = tick;
+      diff = (endTick >> 0) - (startTick >> 0); // Unsigned 32 bit arithmetic
+      console.log(diff / 2 / MICROSECDONDS_PER_CM);
+    }
+  });
+}());
+
+// Trigger a distance measurement one per second
+setInterval(function () {
+  trigger.trigger(10, 1); // Set trigger high for 10 microseconds
+}, 1000);
+